etl/pyproject.toml
[tool.poetry]
name = "etl"
version = "0.1.0"
description = ""
authors = ["Nikita <zigman.nikita@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.10.7"
psycopg2-binary = "^2.9.6"
loguru = "^0.7.0"
python-dotenv = "^1.0.0"
requests = "2.30.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.3.0"
mypy = "^1.3.0"
pre-commit = "^3.3.1"
ipython = "^8.13.2"
types-requests = "^2.30.0.0"
urllib3 = "1.26.15"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
